requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.1
supabase>=2.3.4
python-dotenv>=1.0.1
pydantic>=2.10.0
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
